fastapi/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
Requests==2.31.0
mysql-connector-python==8.0.33
python-jose==3.3.0